Modularizacion inicial de clase personaje y agregado de cambio readme
