+++ README.md
@@ -2,3 +2,6 @@
 
 Juego desarrollado por estudiantes de ingenieria en computacion, en donde se realizara un proyecto aplicando metodologia scrum.
 Tambien se usaran diferentes patrones de diseño en nuestro codigo. 
+
+
+[![Ask DeepWiki](https://deepwiki.com/badge.svg)](https://deepwiki.com/JosiasTomasnanez/SatisPlanning-game)
