+++ src/SatisPlanning/personaje.py
@@ -9,7 +9,8 @@ from SatisPlanning.componente_animacion import ComponenteAnimacion
 from SatisPlanning.componente_agregar_objeto import ComponenteAgregarObjeto
 
 class Personaje(Objeto):
-    def __init__(self, x, y, ancho, alto):
+    def __init__(self, x, y, ancho, alto, sprite, dinamico=True, tangible=True):
+        super().__init__(x, y, ancho, alto, sprite, dinamico=dinamico, tangible=tangible)
         """
         Inicializa el personaje con posición, sprites y un inventario.
 
@@ -18,20 +19,22 @@ class Personaje(Objeto):
         :param ancho: Ancho del personaje.
         :param alto: Altura del personaje.
         """
+        '''
+        Esto deberia inciarce en cada sublclase ya que es particular de cada personaje.
+
         super().__init__(x, y, ancho, alto, obtener_ruta_asset("p3.png"), dinamico=True, tangible=True)
+        
+        Esto deberia iniciarce en cada subclase, ya que cada personaje tiene 
         self.vel_x = self.vel_y = 0
         self.en_el_suelo = False
         self.direccion = 1  # 1 para derecha, -1 para izquierda
+        '''
 
-        # Cargar los sprites del personaje
+        # Cargar los sprites del personaje, TAL VEZ DEBERIA MODIFICARSE CUANDO SE AGREGUEN TEXTURAS DE ENEMI
         self.sprites = [
             pygame.image.load(obtener_ruta_asset(f"p{i}.png")) for i in range(1, 8)
         ]
         self.sprites = [pygame.transform.scale(sprite, (ancho, alto)) for sprite in self.sprites]
-
-        # Inventario del personaje
-        self.inventario = Inventario()
-        self.inventario.visible = False
 
         # Componente para manejar la animación
         self.componente_animacion = ComponenteAnimacion(self, self.sprites)
@@ -41,30 +44,6 @@ class Personaje(Objeto):
 
         # Componente para agregar objetos al mundo
         self.componente_agregar_objeto = ComponenteAgregarObjeto(self)
-
-
-
-
-    def manejar_evento(self, evento, mundo):
-        """
-        Maneja eventos relacionados con el personaje, como abrir/cerrar el inventario y seleccionar elementos de la barra rápida.
-        """
-        if evento.type == pygame.KEYDOWN:
-            if evento.key == pygame.K_i:
-                self.inventario.visible = not self.inventario.visible
-
-            if pygame.K_1 <= evento.key <= pygame.K_9:
-                indice_barra = evento.key - pygame.K_1
-                self.inventario.seleccionar_barra_rapida(indice_barra)
-
-            if evento.key == pygame.K_g:
-                item_soltado = self.inventario.soltar_item_seleccionado()
-                if item_soltado:
-                    if(self.direccion == 1):
-                        item_soltado.actualizar_posicion(self.x + 45, self.y - 10)
-                    else:
-                        item_soltado.actualizar_posicion(self.x - 60, self.y - 10)
-                    self.componente_agregar_objeto.agregar_objeto(item_soltado, mundo,True)
 
     def actualizar(self, teclas, mundo):
         """
@@ -81,7 +60,7 @@ class Personaje(Objeto):
         personaje_centrado_y = (camara.alto_pantalla // 2) - (self.alto // 2)
 
         pantalla.blit(self.componente_animacion.imagen_actual, (personaje_centrado_x, personaje_centrado_y))
-        self.inventario.dibujar(pantalla, fuente)
+        #self.inventario.dibujar(pantalla, fuente) 
 
 #el manejo de coliciones y manejo de movimientos o ataques o demas, estaria bueno hacerlo como compoonentes , como usando el patron stratagy,  ya que hay muchos objetos u personajes que van a tener las mismas fisicas de movimiento pero con variables como la gravedad , entre otras diferentes, entonces seria hacer clases que se encarguen de dichos comportamientos, de las cuales los objetos puedan o no hacer uso, mandandole sus caracteristicas, y queda mucho mas modular y extendible el codigo
