+++ src/SatisPlanning/personaje_enemigo.py
@@ -1,0 +1,4 @@
+from .personaje import personaje  
+
+class personaje_enemigo(personaje):
+    pass
